Propuesta v 03-11-11, rev 02

+++ src/Modelo/Estrella.java
@@ -8,7 +8,7 @@ public class Estrella extends ObjetoAstronomicoEsferico {
 
     private double temperatura;
 
-    public Estrella(String nombre, double masa, double diametro, double vRotacion, double temperatura) {
+    public Estrella(String nombre, double masa, double diametro, float vRotacion, double temperatura) {
         super(nombre, masa, diametro, vRotacion);
         this.temperatura = temperatura;
     }
